test(ui): decompose slow component contracts
