+++ docs/internal/development/plans/ui-test-latency/ui-test-coverage-implementation-plan.md
@@ -709,22 +709,34 @@ cause rather than runner alone:
 | Dashboard app-shell lifecycle | `use-dashboard-snapshot-checkpoint-lifecycle.component.test.tsx` | 1,344ms | Replaced by the feature-owned `usePersistedTimelineCheckpoint` Bun contract; focused execution is about 203ms. |
 | Dashboard app-shell session wiring | `dashboard-session-timeline-isolation.component.test.tsx` | 1,089ms | Removed as duplicate coverage of timeline entries, dashboard session state, header tabs, and the snapshot hook. |
 | Module-mocked async hook | `use-dashboard-checkpoint-preflight.test.tsx` | 778ms | Replaced four Vitest files (884 lines) with one feature-owned Bun core contract using explicit resolver and mutation effects; focused execution is about 635ms and no longer constructs mocked application providers. |
-| Omnibus widget rendering | `current-selection-widget.provider-session-selection.test.tsx` | 720ms | Extract the provider-session selector contract from the complete current-selection widget. |
+| Omnibus widget rendering | `current-selection-widget.provider-session-selection.test.tsx` | 720ms | Removed the redundant 500-line widget suite; inference-history selection state now runs in the feature-owned `useSelectedProviderSessionState` Bun contract (about 164ms), while selection controls remain covered by WorkItemCard and workstation-attempt contracts. |
 | Graph/editor host composition | `react-flow-current-activity-card-host-contract.test.tsx` | 700ms | Retain the small host boundary; continue replacing aggregate imports with owner modules. |
-| Cross-owner dashboard wiring | `dashboard-replay-wiring.component.test.tsx` | 697ms | Keep one wiring proof and move projection cases to timeline/dashboard owners. |
-| Graph validation and charts | workflow editor validation and work chart tests | 675–691ms | Split pure validation/projection from rendered controls before Bun migration. |
-| Current-selection editing | prompt-edit save-enablement tests | 670ms | Test the save state machine and focused editor control independently. |
-| Cross-owner trace wiring | `dashboard-trace-wiring.component.test.tsx` | 616ms | Keep only the dashboard-to-trace handoff; move trace rendering cases to the trace feature. |
+| Cross-owner dashboard wiring | `dashboard-replay-wiring.component.test.tsx` | 697ms | Removed the app-shell duplicate: stream append and fixed-tick behavior are owned by the Bun snapshot hook, timeline store, slider, and bento snapshot contracts. |
+| Graph validation and charts | workflow editor validation and work chart tests | 675–691ms | WorkChart's 18 rendered contracts now run under Bun; workflow editor validation remains the next graph target. |
+| Current-selection editing | prompt-edit save-enablement tests | 670ms | Removed the duplicate widget suite; invalid-to-valid prompt recovery and save submission already run in the feature-owned Bun save contract. |
+| Cross-owner trace wiring | `dashboard-trace-wiring.component.test.tsx` | 616ms | Removed the app-shell duplicate after reconciling event lineage, trace merging, trace rendering, and selection callbacks with their timeline, trace, and current-selection owners. |
 
 The dominant categories remain production import fan-out, complete dashboard or
 widget mounts, and process-global module mocks. Assertion execution is not the
 primary cost. Migration therefore starts by narrowing ownership and dependency
 seams; changing only the runner for an app-sized test preserves most latency.
 
-After the preflight migration, the complete component command owns 228 Bun
-files and 130 remaining Vitest files and completes locally in 83.74s: about
-35.29s in Bun and 48.45s in Vitest. The last measured unit lane is 18.10s, so
-the combined non-browser correctness feedback is approximately 101.84s.
+After the current-selection and cross-owner reconciliation, the complete
+component command owns 231 Bun files and 124 remaining Vitest files and
+completes locally in 79.49s: 36.03s in Bun and 43.46s in Vitest. The last
+measured unit lane is 18.50s, so the combined non-browser correctness feedback
+is approximately 97.99s.
+
+The same reconciliation removed the canonical-section omnibus widget loop.
+Every detail owner structurally composes `SelectionDetailLayout`; the base
+layout and individual detail-card contracts own that behavior without mounting
+every selection kind through the aggregate widget.
+
+WorkChart now owns a dedicated `components/work-chart/` package directory.
+Production modules and Storybook coverage stay at that owner root, while the
+rendered legend, series/state, and interaction contracts live under its
+`contracts/` directory with one shared fixture module. This reduced the flat
+work-outcome component folder below its legacy allowance, which was removed.
 
 Keep the taxonomy as unit, component, and browser tests. Native Bun is an
 execution sublane of component tests, not a fourth test type:
