+++ ui/src/features/current-selection/work-selection/hooks/useSelectedProviderSessionState.bun.component.test.tsx
@@ -1,4 +1,5 @@
 import { act, renderHook, waitFor } from "@testing-library/react";
+import { describe, expect, it } from "bun:test";
 import type { DashboardWorkstationRequest } from "../../../../api/dashboard/types";
 import type { CurrentSelectionState } from "../../hooks/core/useCurrentSelection";
 import { useSelectedProviderSessionState } from "./useSelectedProviderSessionState";
