+++ ui/src/features/work-outcome/components/d3-information-card.tsx
@@ -3,8 +3,11 @@ import { DashboardWidgetFrame } from "../../bento/components/dashboard-widget-fr
 import { getDashboardWorkChartSeriesDefinitions } from "../lib/chart-contract";
 import type { WorkChartModel } from "../lib/trends";
 import { getWorkOutcomeMessages } from "../messages/work-outcome";
-import type { WorkChartSeriesDefinition, WorkChartState } from "./work-chart";
-import { WorkChart } from "./work-chart";
+import type {
+  WorkChartSeriesDefinition,
+  WorkChartState,
+} from "./work-chart/work-chart";
+import { WorkChart } from "./work-chart/work-chart";
 
 export interface WorkChartCardProps {
   chartState?: WorkChartState;
